backend/requirements.txt
Django>=5.0,<5.1
djangorestframework
psycopg2-binary
django-cors-headers
gunicorn